requirements.txt
transformers==4.41.2
torch
sentencepiece
python-dotenv
sentence-transformers==2.7.0
numpy==1.26.4